moved text lib from cli to cli-lib and updated usage
