+++ packages/cli/commands/auth.js
@@ -19,6 +19,7 @@ const {
   accountNameExistsInConfig,
   writeConfig,
 } = require('@hubspot/cli-lib/lib/config');
+const { commaSeparatedValues } = require('@hubspot/cli-lib/lib/text');
 const {
   promptUser,
   personalAccessKeyPrompt,
@@ -33,7 +34,6 @@ const {
 } = require('../lib/commonOpts');
 const { logDebugInfo } = require('../lib/debugInfo');
 const { trackCommandUsage } = require('../lib/usageTracking');
-const { commaSeparatedValues } = require('../lib/text');
 const { authenticateWithOauth } = require('../lib/oauth');
 
 const ALLOWED_AUTH_METHODS = [
