+++ packages/cli/commands/create.js
@@ -8,6 +8,7 @@ const { logger } = require('@hubspot/cli-lib/logger');
 const { createProject } = require('@hubspot/cli-lib/projects');
 const { createFunction } = require('@hubspot/cli-lib/functions');
 const { GITHUB_RELEASE_TYPES } = require('@hubspot/cli-lib/lib/constants');
+const { commaSeparatedValues } = require('@hubspot/cli-lib/lib/text');
 const { fetchJsonFromRepository } = require('@hubspot/cli-lib/github');
 
 const { setLogLevel, getAccountId } = require('../lib/commonOpts');
@@ -21,7 +22,6 @@ const {
   createApiSamplePrompt,
   overwriteSamplePrompt,
 } = require('../lib/createApiSamplePrompt');
-const { commaSeparatedValues } = require('../lib/text');
 
 const TYPES = {
   function: 'function',
