+++ packages/cli/lib/validation.js
@@ -10,6 +10,7 @@ const {
   OAUTH_AUTH_METHOD,
   PERSONAL_ACCESS_KEY_AUTH_METHOD,
 } = require('@hubspot/cli-lib/lib/constants');
+const { commaSeparatedValues } = require('@hubspot/cli-lib/lib/text');
 const { getAbsoluteFilePath } = require('@hubspot/cli-lib/path');
 const { getOauthManager } = require('@hubspot/cli-lib/oauth');
 const {
@@ -17,7 +18,6 @@ const {
 } = require('@hubspot/cli-lib/personalAccessKey');
 const { getCwd, getExt } = require('@hubspot/cli-lib/path');
 const { getAccountId, getMode } = require('./commonOpts');
-const { commaSeparatedValues } = require('./text');
 const fs = require('fs');
 const path = require('path');
